IMP-1167: Reflect different account states on Mobile Web Portal. css & jsp changes.

+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/User.java
@@ -522,7 +522,7 @@ public class User implements Serializable {
 	}
 
     public Date getNextSubPaymentAsDate() {
-        return new Date(this.nextSubPayment*1000);
+        return new Date((long)this.nextSubPayment*1000L);
     }
 
 	public void setNextSubPayment(int nextSubPayment) {
